update TaylorGreen2D test case
add compare_results.py, call this in run.sh

+++ setups/TaylorGreen2D/run.sh
@@ -7,19 +7,23 @@ echo run.sh now running on "$HOSTNAME"
 
 source set_env_variables.sh
 
-#-------------------------------------------------
-# prepare things for Taylor-Green vortex test run
-#-------------------------------------------------
-rm -f input
-ln -s setups/TaylorGreen2D input
-make clean
-make flow.x
-make outdirs
+do_run=True
 
-#---------------------------------------------------
-# do the run (it's set up for a 2x2 processor grid)
-#---------------------------------------------------
-"$MPIRUN" -np 4 ./flow.x
+#------------------------------------------------------------
+# prepare things for the TaylorGreen2D test
+#------------------------------------------------------------
+if( "$do_run" ) then
+	rm -f input
+	ln -s setups/TaylorGreen2D input
+	make clean
+	make flow.x
+	make outdirs
+
+	#---------------------------------------------------
+	# do the run (it's set up for a 4x1 processor grid)
+	#---------------------------------------------------
+	"$MPIRUN" -np 4 ./flow.x
+fi
 
 #---------------------------------------------------------------
 # when run is complete, stitch together the output files
@@ -29,6 +33,9 @@ make outdirs
 cd input/data_tools
 "$PYTHON" concatenate_results.py "$FLOW_SOLVE_ROOT"
 "$PYTHON" python_scripts/plot_cfl.py "$FLOW_SOLVE_ROOT"  s    # plot the time scale seconds
+"$PYTHON" python_scripts/compare_results.py "$FLOW_SOLVE_ROOT"
 
 cd "$FLOW_SOLVE_ROOT"
-open -a Preview output/figures/cfl.pdf
+open -a Preview output/figures/cfl.pdf output/figures/comparison.pdf
+
+"$NCVIEW" output/slices/2D/XY* &
